P.O.C 0.05
There is a 'playable' game now. :)
Implemented a count that has no impact at the moment, for the amount of apples each hand is picking from the trees. I also did some minor refactoring to utilise accessors more rather than storing most of the values in the state. This means that cost values, as well as per sec values will be stored in accessors, whereas current values for currencies etc. will be stored in the localStorage state key.